chore: rename user.ts to userModel.ts
