+++ .gitignore
@@ -678,3 +678,4 @@ fabric.properties
 
 .idea/
 .txt
+dump.rdb
