+++ server/src/seeds/user-seeds.ts
@@ -1,4 +1,4 @@
-import { User } from '../models/user.js';
+import { User } from '../models/userModel.js';
 
 export const seedUsers = async () => {
   await User.bulkCreate([
